==> notebooks/step1_data_preprocessing.ipynb <==
# 1. Data Preprocessing

## 1.1 Load Dataset
---
The following code will load in the full dataset from `merged_reviews.csv`.

from google.colab import drive
import os
import pandas as pd

# Mount Google Drive
drive.mount('/content/drive', force_remount=True)

# Define project paths
project_dir = '/content/drive/MyDrive/bt4222_group_6/bt4222_group_6_amazon'
data_dir = '/content/drive/MyDrive/bt4222_group_6/bt4222_group_6_amazon/data'

merged_df = pd.read_csv(os.path.join(data_dir, 'merged_reviews.csv'))

## Amazon US Customers Review Dataset
---
### User Behaviour Data
* `customer_id`: Random identifier that can be used to aggregate reviews written by a single author.
* `review_id`: The unique ID of the review.
* `vine`: Review was written as part of the Vine program.
* `verified_purchase`: The review is on a verified purchase.

### Product Metadata
* `product_id`: The unique Product ID the review pertains to.
* `product_parent`: Random identifier that can be used to aggregate reviews for the same product.
* `product_title`: Title of the product.
* `product_category`: Broad product category that can be used to group reviews

### Review-based Features
* `marketplace: 2 letter country code of the marketplace where the review was written.
* `star_rating`: The 1-5 star rating of the review.
* `helpful_votes`: Number of helpful votes.
* `total_votes`: Number of total votes the review received.
* `review_headline`: The title of the review.
* `review_body`: The review text.
* `review_date`: The date the review was written.

## 1.2 Inspect Dataset
---
The following code aims to gather more insights into the structure of the `merged_reviews.csv` dataset to prepare for Data Cleaning.

merged_df.info()

merged_df.describe()

merged_df.head()

merged_df.isnull().sum()

## 1.3 Data Cleaning
---
The following code aims to handle missing rows and irrelevant columns such as `marketplace` and `review_id`.

# Drop columns you don't need
cleaned_df = merged_df.drop(columns=['marketplace', 'review_id'])

# Drop rows with missing values in key fields like review_body
cleaned_df = cleaned_df.dropna(subset=['review_body', 'star_rating'])

# Fill missing values to keep certain rows
cleaned_df['product_title'] = cleaned_df['product_title'].fillna('Unknown Product')
cleaned_df['review_headline'] = cleaned_df['review_headline'].fillna('')
cleaned_df['review_date'] = pd.to_datetime(cleaned_df['review_date'], errors='coerce')
cleaned_df['review_date'] = cleaned_df['review_date'].fillna(pd.Timestamp('1970-01-01'))

## 1.4 Filtering
---
The following code will filter `cleaned_df` for data from the **last 4 years** and for users who have made **5 or more purchases**, and ensure that the reviews are verified, resulting in `filtered_reviews.csv`. Note that there is the assumption that leaving a review implies that a purchase was made.

### 1.4.1 Filter for verified data (verified_purchase == 1)

filtered_df = cleaned_df.copy()

filtered_df = filtered_df[filtered_df['verified_purchase'] == 'Y']

### 1.4.2 Filter for data from the last 4 years

# Convert review_date column to datetime
filtered_df['review_date'] = pd.to_datetime(filtered_df['review_date'], errors='coerce')

# Sort by review_date in descending order (latest first)
filtered_df = filtered_df.sort_values(by="review_date", ascending=False)

# Get the latest review date
latest_date = filtered_df["review_date"].max()

# Filter for reviews within the last 4 years
four_years_ago = latest_date - pd.DateOffset(years=4)
filtered_df = filtered_df[filtered_df["review_date"] >= four_years_ago]

### 1.4.3 Filter for users with 5 or more purchases

# Filter for users with 5 or more reviews (purchases)
user_counts = filtered_df['customer_id'].value_counts()
eligible_users = user_counts[user_counts >= 5].index
filtered_df = filtered_df[filtered_df['customer_id'].isin(eligible_users)]

filtered_df.to_csv(os.path.join(data_dir, 'filtered_reviews.csv'),
    index=False
)

filtered_df.info()

filtered_df.shape

filtered_df.head()
